ref: split IERC1155Receiver interface to another file
